backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-multipart==0.0.6
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
torch==2.1.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
xgboost==2.0.3
catboost==1.2.2
optuna==3.14.0
chardet==5.2.0
google-genai==0.4.0
feature-engine==1.6.1
openpyxl==3.11.0
tqdm==4.66.1
